Static page contents may be configured outside the source code.
The system now supports database-driven page contents for the following
"static" pages:

* Attribution Scholarship (`scholarship`)
* History (`history`)
* Contact Us (`contact`)
* Related Sites (`sites`)

In order to support this, a new database table called `page_contents`
has been created according to the following schema:
```
+------------+--------------+------+-----+---------+----------------+
| Field      | Type         | Null | Key | Default | Extra          |
+------------+--------------+------+-----+---------+----------------+
| id         | int(11)      | NO   | PRI | NULL    | auto_increment |
| page_key   | varchar(255) | YES  |     | NULL    |                |
| html       | text         | YES  |     | NULL    |                |
| created_at | datetime     | NO   |     | NULL    |                |
| updated_at | datetime     | NO   |     | NULL    |                |
+------------+--------------+------+-----+---------+----------------+
```

Of these fields, the meaningful ones are
* `page_key` is the key to a page in the URL.  Currently valid page keys
are `scholarship`, `history`, `contact`, and `sites`, however this can be
expanded flexibly.
* `html` is the HTML source code for the page.

If no page contents record is found for an existing page key, the
static text in the Github repository for that page is used instead.